pyproject.toml
[project]
name = "big0-dev-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "babel>=2.17.0",
    "bs4>=0.0.2",
    "jinja2>=3.1.6",
    "jsmin>=3.0.1",
    "markdown>=3.7",
    "minify-html>=0.16.4",
    "pyyaml>=6.0.2",
    "rcssmin>=1.2.1",
    # SEO Engine dependencies
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-analytics-data>=0.18.0",
    "pytrends>=4.9.0",
]
